+++ inst/testme/test-progressify-crossmap-future.R
@@ -11,18 +11,18 @@ fcn <- function(x, y) x * y
 
 
 ## -------------------------------------------------------
-## xmap family
+## future_xmap family
 ## -------------------------------------------------------
 exprs <- list(
-  xmap     = quote(xmap(xs, fcn)),
-  xmap     = quote(crossmap::xmap(xs, fcn)),
-  xmap     = quote(crossmap::xmap(.l = xs, .f = fcn)),
-  xmap_dbl = quote(xmap_dbl(xs, ~ .y * .x)),
-  xmap_dbl = quote(crossmap::xmap_dbl(xs, ~ .y * .x)),
-  xmap_int = quote(xmap_int(xs, ~ .y * .x)),
-  xmap_chr = quote(xmap_chr(xs, ~ paste(.y, "*", .x, "=", .y * .x))),
-  xmap_lgl = quote(xmap_lgl(xs, ~ .y > .x)),
-  xmap_vec = quote(xmap_vec(xs, fcn))
+  future_xmap     = quote(future_xmap(xs, fcn)),
+  future_xmap     = quote(crossmap::future_xmap(xs, fcn)),
+  future_xmap     = quote(crossmap::future_xmap(.l = xs, .f = fcn)),
+  future_xmap_dbl = quote(future_xmap_dbl(xs, ~ .y * .x)),
+  future_xmap_dbl = quote(crossmap::future_xmap_dbl(xs, ~ .y * .x)),
+  future_xmap_int = quote(future_xmap_int(xs, ~ .y * .x)),
+  future_xmap_chr = quote(future_xmap_chr(xs, ~ paste(.y, "*", .x, "=", .y * .x))),
+  future_xmap_lgl = quote(future_xmap_lgl(xs, ~ .y > .x)),
+  future_xmap_vec = quote(future_xmap_vec(xs, fcn))
 )
 
 for (kk in seq_along(exprs)) {
@@ -61,12 +61,12 @@ for (kk in seq_along(exprs)) {
 
 
 ## -------------------------------------------------------
-## xwalk
+## future_xwalk
 ## -------------------------------------------------------
 res_walk <- list()
 truth_walk <- list()
 
-expr <- quote(xwalk(xs, function(x, y) { list(x = x, y = y) }))
+expr <- quote(future_xwalk(xs, function(x, y) { list(x = x, y = y) }))
 truth <- eval(expr)
 res_walk_truth <- res_walk
 
@@ -77,12 +77,12 @@ stopifnot(identical(res_walk, res_walk_truth))
 
 
 ## -------------------------------------------------------
-## xmap_mat, xmap_arr
+## future_xmap_mat, future_xmap_arr
 ## -------------------------------------------------------
 exprs <- list(
-  xmap_mat = quote(xmap_mat(xs, fcn)),
-  xmap_mat = quote(crossmap::xmap_mat(.l = xs, .f = fcn)),
-  xmap_arr = quote(xmap_arr(xs, fcn))
+  future_xmap_mat = quote(future_xmap_mat(xs, fcn)),
+  future_xmap_mat = quote(crossmap::future_xmap_mat(.l = xs, .f = fcn)),
+  future_xmap_arr = quote(future_xmap_arr(xs, fcn))
 )
 
 for (kk in seq_along(exprs)) {
@@ -121,20 +121,20 @@ for (kk in seq_along(exprs)) {
 
 
 ## -------------------------------------------------------
-## purrr-extension *_vec functions
+## future purrr-extension *_vec functions
 ## -------------------------------------------------------
 ys <- 1:5
 fcn2 <- function(x) x^2
 
 exprs <- list(
-  map_vec  = quote(map_vec(ys, fcn2)),
-  map_vec  = quote(crossmap::map_vec(ys, fcn2)),
-  map2_vec = quote(map2_vec(ys, ys, fcn)),
-  map2_vec = quote(crossmap::map2_vec(ys, ys, fcn)),
-  pmap_vec = quote(pmap_vec(list(ys, ys), fcn)),
-  pmap_vec = quote(crossmap::pmap_vec(list(ys, ys), fcn)),
-  imap_vec = quote(imap_vec(ys, ~ .x + .y)),
-  imap_vec = quote(crossmap::imap_vec(ys, ~ .x + .y))
+  future_map_vec  = quote(future_map_vec(ys, fcn2)),
+  future_map_vec  = quote(crossmap::future_map_vec(ys, fcn2)),
+  future_map2_vec = quote(future_map2_vec(ys, ys, fcn)),
+  future_map2_vec = quote(crossmap::future_map2_vec(ys, ys, fcn)),
+  future_pmap_vec = quote(future_pmap_vec(list(ys, ys), fcn)),
+  future_pmap_vec = quote(crossmap::future_pmap_vec(list(ys, ys), fcn)),
+  future_imap_vec = quote(future_imap_vec(ys, ~ .x + .y)),
+  future_imap_vec = quote(crossmap::future_imap_vec(ys, ~ .x + .y))
 )
 
 for (kk in seq_along(exprs)) {